Change 1/3 - Move Powerup from extending Mob
Previously, the Powerup class extended from the Mob class just to
inherit the move method in Mob. However, it also inherited the
destructible interface even though the Powerup class is not a
destructible entity, as well as the blockedX and blockedY attributes
that Powerup never uses.

Powerup now extends Entity directly instead of Mob.  Added move method
to Powerup class that performs the actions it did when using move from
Mob but, without the unneeded changes to the blockedX and blockedY
attributes.

Created Moveable interface to make all entity objects that need to move
implement this interface.  This is because not all entities use the move
method inherited from Entity, including Powerup.  The move method was
removed from the Entity class.  All classes that implemented move and
inherit from Entity now implement the Moveable interface, including
PowerUp.